[feat] 배치 테스트
