+++ build.gradle
@@ -42,6 +42,7 @@ dependencies {
 	implementation 'com.fasterxml.jackson.module:jackson-module-kotlin'
 	implementation 'org.jetbrains.kotlin:kotlin-reflect'
 	implementation 'org.springframework.boot:spring-boot-starter-security'
+	implementation 'org.springframework.boot:spring-boot-starter-jdbc'
 
 	// Lombok
 	compileOnly 'org.projectlombok:lombok'
@@ -97,9 +98,6 @@ dependencies {
 	// Spring Batch
 	implementation 'org.springframework.boot:spring-boot-starter-batch'
 	testImplementation 'org.springframework.batch:spring-batch-test'
-
-	// Exposed
-	implementation 'org.jetbrains.exposed:exposed-spring-boot-starter:0.48.0'
 }
 
 tasks.withType(KotlinCompile) {
