+++ src/main/kotlin/com/meokq/api/batch/scheduler/Scheduler.kt
@@ -1,26 +1,30 @@
 package com.meokq.api.batch.scheduler
 
-import com.meokq.api.batch.controller.BatchController
-import com.meokq.api.user.enums.UserStatus
-import org.springframework.batch.core.JobParameters
+import com.meokq.api.batch.job.BatchJob
 import org.springframework.batch.core.JobParametersBuilder
 import org.springframework.batch.core.launch.JobLauncher
 import org.springframework.scheduling.annotation.Scheduled
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
 import java.sql.Date
 import java.time.LocalDate
 
+@RestController
+@RequestMapping("/api")
+class Scheduler(
+    private val batchJob: BatchJob,
+    private val jobLauncher : JobLauncher){
 
-class Scheduler(val batchController: BatchController,val customerRepository: C,val jobLauncher : JobLauncher){
-    @Scheduled(cron="0 0 24 * * ?")
+    @Scheduled(cron="0 0 0 * * ?")
+    @PostMapping(value = ["/open/test"])
     fun run() {
-        val CHUNK_SIZE = 1000
-        val dataSize = customerRepository.countByStatus(UserStatus.DORMANT)
-        val iteratorSize = if (dataSize > CHUNK_SIZE) dataSize / CHUNK_SIZE else 1
-
-        val jobExecution = jobLauncher.run(batchController.updateCustomerStatusJob(),
-            JobParametersBuilder().addDate("date",Date.valueOf(LocalDate.now()))
+        println(LocalDate.now().toString())
+        val jobExecution = jobLauncher.run(batchJob.updateCustomerStatusJob(),
+            JobParametersBuilder().addString("date",LocalDate.now().toString())
                 .toJobParameters()
             )
+
 
     }
 }
