+++ src/main/kotlin/com/meokq/api/batch/step/StepService.kt
@@ -1,10 +1,9 @@
-package com.meokq.api.batch.service
+package com.meokq.api.batch.step
 
 import org.springframework.batch.core.Step
 import org.springframework.beans.factory.annotation.Value
 import java.time.LocalDate
 
 interface StepService {
-    fun updateCustomerStatusStep(@Value("#{jobParameters='date'}") date : LocalDate): Step
-
+    fun updateCustomerStatusStep(): Step
 }
