+++ src/main/kotlin/com/meokq/api/core/config/SecureConfig.kt
@@ -39,6 +39,9 @@ class SecureConfig(
             //.requestMatchers("/api/**").authenticated()
             //.requestMatchers("/auth/**").permitAll()
             .requestMatchers("/h2-console/**").permitAll()
+//            .requestMatchers("/api/admin").hasRole("ADMIN")
+//            .requestMatchers("/api/boss").hasRole("BOSS")
+//            .requestMatchers("/api/customer").hasRole("CUSTOMER")
             .anyRequest().permitAll()
             .and()
             .headers().frameOptions().sameOrigin()
