+++ src/main/kotlin/com/meokq/api/user/model/Boss.kt
@@ -7,6 +7,7 @@ import com.meokq.api.user.enums.UserStatus
 import jakarta.persistence.*
 import org.hibernate.annotations.UuidGenerator
 import org.jetbrains.annotations.NotNull
+import java.time.LocalDate
 
 @Entity(name = "tb_boss")
 data class Boss(
@@ -20,9 +21,12 @@ data class Boss(
     var email : String? = null,
     @Enumerated(EnumType.STRING)
     var channel: AuthChannel? = null,
+    @Column(name = "withdrawn_at")
+    var withdrawnAt : LocalDate? = null
 ) : BaseModel(){
     constructor(request : LoginReq) : this(
         email = request.email,
         channel = request.channel
     )
+
 }
