+++ src/main/kotlin/com/meokq/api/user/service/BossService.kt
@@ -12,6 +12,7 @@ import com.meokq.api.user.response.UserResp
 import com.meokq.api.user.response.WithdrawResp
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Service
+import java.time.LocalDate
 
 @Service
 class BossService (
@@ -42,6 +43,7 @@ class BossService (
         try {
             val model = findModelById(userId)
             model.status = model.status.withdrawnAction()
+            model.withdrawnAt = LocalDate.now()
             val result = saveModel(model)
             return WithdrawResp(result)
