+++ src/main/kotlin/com/meokq/api/user/service/CustomerService.kt
@@ -18,7 +18,7 @@ import com.meokq.api.user.response.UserResp
 import com.meokq.api.user.response.WithdrawResp
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Service
-import java.time.LocalDateTime
+import java.time.LocalDate
 
 @Service
 class CustomerService(
@@ -88,7 +88,7 @@ class CustomerService(
         try {
             val model = findModelById(userId)
             model.status = model.status.withdrawnAction()
-            model.withdrawnAt = LocalDateTime.now()
+            model.withdrawnAt = LocalDate.now()
             val result = saveModel(model)
             return WithdrawResp(result)
