+++ src/main/resources/application-local.yml
@@ -1,6 +1,7 @@
 server:
   port: 9090
 
+
 spring:
   config:
     import: common.yml
@@ -8,7 +9,7 @@ spring:
   # H2 Database
   datasource:
     driver-class-name: org.h2.Driver
-    url: jdbc:h2:mem:test   # H2 DB (In-Memory Mode)
+    url: jdbc:h2:mem:db;MODE=MYSQL;   # H2 DB (In-Memory Mode)
     username: sa
     password:
 
@@ -29,6 +30,11 @@ spring:
         #format_sql: true
         #show_sql: true
     defer-datasource-initialization: true
+
+  batch:
+    jdbc:
+      initialize-schema: always
+
 
 logging:
   level:
