+++ src/main/resources/data.sql
@@ -38,15 +38,13 @@ VALUES
     ('IMMA2024031115000583', 'RECEIPT', 3072);
 
 -- Insert data into tb_customer table
-INSERT INTO tb_customer (customer_id, status, email, nickname)
+INSERT INTO `tb_customer` (`customer_id`, `create_date`, `update_date`, `channel`, `email`, `nickname`, `status`, `withdrawn_at`)
 VALUES
-    ('CS10000001', 'ACTIVE', 'user1@example.com', 'nickname1'),
-    ('CS10000002', 'DORMANT', 'user2@example.com', 'nickname2'),
-    ('CS10000003', 'WITHDRAWN', 'user3@example.com', 'nickname3'),
-    ('CS10000004', 'WITHDRAWN', 'user4@example.com', 'nickname3'),
-    ('CS10000005', 'WITHDRAWN', 'user5@example.com', 'nickname3'),
-    ('CS10000006', 'WITHDRAWN', 'user6@example.com', 'nickname3'),
-    ('CS10000007', 'WITHDRAWN', 'user7@example.com', 'nickname3');
+    ('110804aa-a3f9-4894-93d9-9b446e583b27', '2024-01-30 12:43:05.796058', '2024-01-30 12:43:05.796078', 'APPLE', 'dsss@sample.com', 'USER2024013012430596', 'ACTIVE', '2024-01-30'),
+    ('4e7b25a2-5c65-4de4-82f7-f0034f5d4615', '2024-02-08 15:10:39.177130', '2024-02-10 23:32:44.243646', 'APPLE', 'dlwlsgml8847@sample.com', 'USER2024020815103834', 'ACTIVE', '2024-01-30'),
+    ('82eb81c2-7df9-4e47-9362-c71c6ac78f60', '2024-02-14 14:33:10.481530', '2024-02-14 14:33:10.481567', 'APPLE', 'wlsgml8847@sample.com', 'USER2024021414331002', 'ACTIVE', '2024-01-30'),
+    ('f6744202-f40f-4ce7-b00f-1a8d10456454', '2024-03-05 14:12:32.579302', '2024-03-05 14:12:32.579316', 'KAKAO', 'email@sample.com', 'USER2024030514123269', 'ACTIVE', '2024-01-30'),
+    ('feeb066f-a118-4dfd-a141-eb8d6f31b8b1', '2024-02-05 13:11:35.501047', '2024-02-05 13:11:35.501083', 'APPLE', '8847wlsgml@naver.com', 'USER2024020513113554', 'ACTIVE', '2024-01-30');
 
 -- Insert data into tb_quest table
 INSERT INTO tb_quest (quest_id, status, market_id)
