--- src/test/kotlin/com/meokq/api/batch/service/batchServiceTest.kt
@@ -1,11 +1,0 @@
-package com.meokq.api.batch.service
-
-import org.junit.jupiter.api.Assertions.*
-import org.junit.jupiter.api.Test
-
-class batchServiceTest{
-    @Test
-    fun test(){
-        assertEquals(1, 1)
-    }
-}
